LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/play-animation-expected.txt


PASS Playing an animations aligns the start time with the start of the active range
PASS Playing an animations with a negative playback rate aligns the start time with the end of the active range
PASS Start time set while play pending is preserved.
PASS Current time set while play pending is preserved.
FAIL Playing a running animation resets a sticky start time assert_approx_equals: values do not match for "undefined" expected 0 +/- 0.125 but got 10
FAIL Playing a finished animation restarts the animation aligned at the start assert_approx_equals: values do not match for "undefined" expected 0 +/- 0.125 but got 100
FAIL Playing a finished and reversed animation restarts the animation aligned at the end assert_approx_equals: values do not match for "undefined" expected 100 +/- 0.125 but got 0
FAIL Playing a pause-pending but previously finished animation realigns with the scroll position assert_approx_equals: values do not match for "After aborting a pause when finished, the start time should no longer be sticky" expected 0 +/- 0.125 but got -50
FAIL Playing a finished animation clears the start time assert_approx_equals: values do not match for "start time is zero" expected 0 +/- 0.125 but got -100
PASS The ready promise should be replaced if the animation is not already pending
PASS A pending ready promise should be resolved and not replaced when the animation enters the running state
FAIL Resuming an animation from paused realigns with scroll position. assert_approx_equals: values do not match for "undefined" expected 0 +/- 0.125 but got -50
PASS If a pause operation is interrupted, the ready promise is reused
FAIL A pending playback rate is used when determining timeline range alignment assert_approx_equals: values do not match for "undefined" expected 100 +/- 0.125 but got -100
PASS Playing a canceled animation sets the start time
PASS Playing a canceled animation backwards sets the start time
